refactor methods crud pessoa

+++ src/main/java/com/projetointegrador/app/rest/Repo/CarroRepository.java
@@ -4,7 +4,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.projetointegrador.app.rest.Models.Carro;
 
-public interface CarroRepository extends JpaRepository<Carro, Long> {
+public interface CarroRepository extends JpaRepository<Carro, Integer> {
 
 }
